+++ mrstorage/sqlizer.go
@@ -1,0 +1,5 @@
+package mrstorage
+
+type Sqlizer interface {
+    ToSql() (string, []interface{}, error)
+}